# src/swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "description" : "sayori",
    "url" : "https://acow-api-dev.sayori.link"
  } ],
  "info" : {
    "description" : "This is a simple API",
    "version" : "1.0.0",
    "title" : "ACOW API",
    "contact" : {
      "email" : "you@your-company.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths" : {
...
    "/users/me/signup-form" : {
      "get" : {
        "summary" : "Retrieve current user's signup form",
        "tags" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-examples" : {
                    "Example 1" : {
                      "success" : true,
                      "form" : {
                        "id" : "7",
                        "user" : "308152268249628672",
                        "repServer" : "Frieren",
                        "anilistUrl" : "",
                        "preferredMedium" : "",
                        "acceptingMedium" : "",
                        "preferredGenres" : "",
                        "bannedGenres" : "",
                        "extremeSpecialParticipation" : false,
                        "sponsorParticipation" : false,
                        "aidParadeParticipation" : false,
                        "gameProfileUrl" : null,
                        "pcPower" : "POTATO",
                        "hasXboxGamePass" : null,
                        "preferredGameGenres" : [ "FIGHTING", "DATING" ],
                        "challengeLevelPreference" : null,
                        "gameLengthPreference" : null,
                        "competitiveBlitzParticipation" : false,
                        "casualBlitzParticipation" : false,
                        "bookClubParticipation" : false,
                        "gameClubParticipation" : false,
                        "notesForStaff" : "",
                        "notesForContractor" : "",
                        "createdAt" : "2026-05-07T20:56:38.000Z",
                        "updatedAt" : "2026-05-07T21:53:17.000Z"
                      }
                    }
                  },
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "form" : {
                      "$ref" : "#/components/schemas/SignUpForm"
                    }
                  }
                },
                "examples" : {
                  "Example 1" : {
                    "value" : {
                      "success" : true,
                      "form" : {
                        "id" : "7",
                        "user" : "308152268249628672",
                        "repServer" : "Frieren",
                        "anilistUrl" : "",
                        "preferredMedium" : "",
                        "acceptingMedium" : "",
                        "preferredGenres" : "",
                        "bannedGenres" : "",
                        "extremeSpecialParticipation" : false,
                        "sponsorParticipation" : false,
                        "aidParadeParticipation" : false,
                        "gameProfileUrl" : null,
                        "pcPower" : "POTATO",
                        "hasXboxGamePass" : null,
                        "preferredGameGenres" : [ "FIGHTING", "DATING" ],
                        "challengeLevelPreference" : null,
                        "gameLengthPreference" : null,
                        "competitiveBlitzParticipation" : false,
                        "casualBlitzParticipation" : false,
                        "bookClubParticipation" : false,
                        "gameClubParticipation" : false,
                        "notesForStaff" : "",
                        "notesForContractor" : "",
                        "createdAt" : "2026-05-07T20:56:38.000Z",
                        "updatedAt" : "2026-05-07T21:53:17.000Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "operationId" : "get-users-signup-template-me"
      },
      "patch" : {
        "summary" : "Update current user's signup form",
        "tags" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {
                      "type" : "string",
                      "x-stoplight" : {
                        "id" : "a13xzenzop3mz"
                      }
                    },
                    "form" : {
                      "$ref" : "#/components/schemas/SignUpForm"
                    }
                  }
                },
                "examples" : {
                  "Example 1" : {
                    "value" : {
                      "success" : true,
                      "form" : {
                        "id" : "7",
                        "user" : "308152268249628672",
                        "repServer" : "Frieren",
                        "anilistUrl" : "",
                        "preferredMedium" : "",
                        "acceptingMedium" : "",
                        "preferredGenres" : "",
                        "bannedGenres" : "",
                        "extremeSpecialParticipation" : false,
                        "sponsorParticipation" : false,
                        "aidParadeParticipation" : false,
                        "gameProfileUrl" : null,
                        "pcPower" : "POTATO",
                        "hasXboxGamePass" : null,
                        "preferredGameGenres" : [ "FIGHTING", "DATING" ],
                        "challengeLevelPreference" : null,
                        "gameLengthPreference" : null,
                        "competitiveBlitzParticipation" : false,
                        "casualBlitzParticipation" : false,
                        "bookClubParticipation" : false,
                        "gameClubParticipation" : false,
                        "notesForStaff" : "",
                        "notesForContractor" : "",
                        "createdAt" : "2026-05-07T20:56:38.000Z",
                        "updatedAt" : "2026-05-07T21:53:17.000Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "operationId" : "patch-users-signup-template-me",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SignUpFormEditable"
              },
              "examples" : {
                "Example 1" : {
                  "value" : {
                    "repServer" : "Frieren",
                    "anilistUrl" : "",
                    "preferredMedium" : "",
                    "acceptingMedium" : "",
                    "preferredGenres" : "",
                    "bannedGenres" : "",
                    "extremeSpecialParticipation" : false,
                    "sponsorParticipation" : false,
                    "aidParadeParticipation" : false,
                    "gameProfileUrl" : null,
                    "pcPower" : "POTATO",
                    "hasXboxGamePass" : null,
                    "preferredGameGenres" : [ "FIGHTING", "DATING" ],
                    "challengeLevelPreference" : null,
                    "gameLengthPreference" : null,
                    "competitiveBlitzParticipation" : false,
                    "casualBlitzParticipation" : false,
                    "bookClubParticipation" : false,
                    "gameClubParticipation" : false,
                    "notesForStaff" : "",
                    "notesForContractor" : ""
                  }
                }
              }
            }
          }
        }
      },
      "parameters" : [ ]
    },
    "/users/{userId}/signup-form" : {
      "get" : {
        "summary" : "[Admin] Retrieve user's signup form",
        "tags" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-examples" : {
                    "Example 1" : {
                      "success" : true,
                      "form" : {
                        "id" : "7",
                        "user" : "308152268249628672",
                        "repServer" : "Frieren",
                        "anilistUrl" : "",
                        "preferredMedium" : "",
                        "acceptingMedium" : "",
                        "preferredGenres" : "",
                        "bannedGenres" : "",
                        "extremeSpecialParticipation" : false,
                        "sponsorParticipation" : false,
                        "aidParadeParticipation" : false,
                        "gameProfileUrl" : null,
                        "pcPower" : "POTATO",
                        "hasXboxGamePass" : null,
                        "preferredGameGenres" : [ "FIGHTING", "DATING" ],
                        "challengeLevelPreference" : null,
                        "gameLengthPreference" : null,
                        "competitiveBlitzParticipation" : false,
                        "casualBlitzParticipation" : false,
                        "bookClubParticipation" : false,
                        "gameClubParticipation" : false,
                        "notesForStaff" : "",
                        "notesForContractor" : "",
                        "createdAt" : "2026-05-07T20:56:38.000Z",
                        "updatedAt" : "2026-05-07T21:53:17.000Z"
                      }
                    }
                  },
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "form" : {
                      "$ref" : "#/components/schemas/SignUpForm"
                    }
                  }
                },
                "examples" : {
                  "Example 1" : {
                    "value" : {
                      "success" : true,
                      "form" : {
                        "id" : "7",
                        "user" : "308152268249628672",
                        "repServer" : "Frieren",
                        "anilistUrl" : "",
                        "preferredMedium" : "",
                        "acceptingMedium" : "",
                        "preferredGenres" : "",
                        "bannedGenres" : "",
                        "extremeSpecialParticipation" : false,
                        "sponsorParticipation" : false,
                        "aidParadeParticipation" : false,
                        "gameProfileUrl" : null,
                        "pcPower" : "POTATO",
                        "hasXboxGamePass" : null,
                        "preferredGameGenres" : [ "FIGHTING", "DATING" ],
                        "challengeLevelPreference" : null,
                        "gameLengthPreference" : null,
                        "competitiveBlitzParticipation" : false,
                        "casualBlitzParticipation" : false,
                        "bookClubParticipation" : false,
                        "gameClubParticipation" : false,
                        "notesForStaff" : "",
                        "notesForContractor" : "",
                        "createdAt" : "2026-05-07T20:56:38.000Z",
                        "updatedAt" : "2026-05-07T21:53:17.000Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "operationId" : "get-users-signup-template"
      },
      "patch" : {
        "summary" : "[Admin] Update user's signup form",
        "tags" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {
                      "type" : "string"
                    },
                    "form" : {
                      "$ref" : "#/components/schemas/SignUpForm"
                    }
                  }
                },
                "examples" : {
                  "Example 1" : {
                    "value" : {
                      "success" : true,
                      "form" : {
                        "id" : "7",
                        "user" : "308152268249628672",
                        "repServer" : "Frieren",
                        "anilistUrl" : "",
                        "preferredMedium" : "",
                        "acceptingMedium" : "",
                        "preferredGenres" : "",
                        "bannedGenres" : "",
                        "extremeSpecialParticipation" : false,
                        "sponsorParticipation" : false,
                        "aidParadeParticipation" : false,
                        "gameProfileUrl" : null,
                        "pcPower" : "POTATO",
                        "hasXboxGamePass" : null,
                        "preferredGameGenres" : [ "FIGHTING", "DATING" ],
                        "challengeLevelPreference" : null,
                        "gameLengthPreference" : null,
                        "competitiveBlitzParticipation" : false,
                        "casualBlitzParticipation" : false,
                        "bookClubParticipation" : false,
                        "gameClubParticipation" : false,
                        "notesForStaff" : "",
                        "notesForContractor" : "",
                        "createdAt" : "2026-05-07T20:56:38.000Z",
                        "updatedAt" : "2026-05-07T21:53:17.000Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "operationId" : "patch-users-signup-template",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SignUpFormEditable"
              },
              "examples" : {
                "Example 1" : {
                  "value" : {
                    "repServer" : "Frieren",
                    "anilistUrl" : "",
                    "preferredMedium" : "",
                    "acceptingMedium" : "",
                    "preferredGenres" : "",
                    "bannedGenres" : "",
                    "extremeSpecialParticipation" : false,
                    "sponsorParticipation" : false,
                    "aidParadeParticipation" : false,
                    "gameProfileUrl" : null,
                    "pcPower" : "POTATO",
                    "hasXboxGamePass" : null,
                    "preferredGameGenres" : [ "FIGHTING", "DATING" ],
                    "challengeLevelPreference" : null,
                    "gameLengthPreference" : null,
                    "competitiveBlitzParticipation" : false,
                    "casualBlitzParticipation" : false,
                    "bookClubParticipation" : false,
                    "gameClubParticipation" : false,
                    "notesForStaff" : "",
                    "notesForContractor" : ""
                  }
                }
              }
            }
          }
        }
      },
      "parameters" : [ {
        "schema" : {
          "type" : "string"
        },
        "name" : "userId",
        "in" : "path",
        "required" : true,
        "description" : "User id associated with the form"
      } ]
    },
...
  },
  "components" : {
    "schemas" : {
...
      "SignUpForm" : {
        "type" : "object",
        "x-examples" : {
          "Example 1" : {
            "id" : "7",
            "user" : "308152268249628672",
            "repServer" : "Frieren",
            "anilistUrl" : "",
            "preferredMedium" : "",
            "acceptingMedium" : "",
            "preferredGenres" : "",
            "bannedGenres" : "",
            "extremeSpecialParticipation" : false,
            "sponsorParticipation" : false,
            "aidParadeParticipation" : false,
            "gameProfileUrl" : null,
            "pcPower" : "POTATO",
            "hasXboxGamePass" : null,
            "preferredGameGenres" : [ "FIGHTING", "DATING" ],
            "challengeLevelPreference" : null,
            "gameLengthPreference" : null,
            "competitiveBlitzParticipation" : false,
            "casualBlitzParticipation" : false,
            "bookClubParticipation" : false,
            "gameClubParticipation" : false,
            "notesForStaff" : "",
            "notesForContractor" : "",
            "createdAt" : "2026-05-07T20:56:38.000Z",
            "updatedAt" : "2026-05-07T21:53:17.000Z"
          }
        },
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "user" : {
            "type" : "string"
          },
          "repServer" : {
            "type" : "string"
          },
          "anilistUrl" : {
            "type" : "string"
          },
          "preferredMedium" : {
            "type" : "string"
          },
          "acceptingMedium" : {
            "type" : "string"
          },
          "preferredGenres" : {
            "type" : "string"
          },
          "bannedGenres" : {
            "type" : "string"
          },
          "extremeSpecialParticipation" : {
            "type" : "boolean"
          },
          "sponsorParticipation" : {
            "type" : "boolean"
          },
          "aidParadeParticipation" : {
            "type" : "boolean"
          },
          "gameProfileUrl" : {
            "type" : "boolean"
          },
          "pcPower" : {
            "type" : "string"
          },
          "hasXboxGamePass" : {
            "type" : "boolean",
            "nullable" : true
          },
          "preferredGameGenres" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "challengeLevelPreference" : {
            "type" : "boolean",
            "nullable" : true
          },
          "gameLengthPreference" : {
            "type" : "boolean",
            "nullable" : true
          },
          "competitiveBlitzParticipation" : {
            "type" : "boolean"
          },
          "casualBlitzParticipation" : {
            "type" : "boolean"
          },
          "bookClubParticipation" : {
            "type" : "boolean"
          },
          "gameClubParticipation" : {
            "type" : "boolean"
          },
          "notesForStaff" : {
            "type" : "string"
          },
          "notesForContractor" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "updatedAt" : {
            "type" : "string"
          }
        }
      },
      "SignUpFormEditable" : {
        "type" : "object",
        "x-examples" : {
          "Example 1" : {
            "repServer" : "Frieren",
            "anilistUrl" : "",
            "preferredMedium" : "",
            "acceptingMedium" : "",
            "preferredGenres" : "",
            "bannedGenres" : "",
            "extremeSpecialParticipation" : false,
            "sponsorParticipation" : false,
            "aidParadeParticipation" : false,
            "gameProfileUrl" : null,
            "pcPower" : "POTATO",
            "hasXboxGamePass" : null,
            "preferredGameGenres" : [ "FIGHTING", "DATING" ],
            "challengeLevelPreference" : null,
            "gameLengthPreference" : null,
            "competitiveBlitzParticipation" : false,
            "casualBlitzParticipation" : false,
            "bookClubParticipation" : false,
            "gameClubParticipation" : false,
            "notesForStaff" : "",
            "notesForContractor" : ""
          }
        },
        "title" : "SignUpFormEditable",
        "properties" : {
          "repServer" : {
            "type" : "string"
          },
          "anilistUrl" : {
            "type" : "string"
          },
          "preferredMedium" : {
            "type" : "string"
          },
          "acceptingMedium" : {
            "type" : "string"
          },
          "preferredGenres" : {
            "type" : "string"
          },
          "bannedGenres" : {
            "type" : "string"
          },
          "extremeSpecialParticipation" : {
            "type" : "boolean"
          },
          "sponsorParticipation" : {
            "type" : "boolean"
          },
          "aidParadeParticipation" : {
            "type" : "boolean"
          },
          "gameProfileUrl" : {
            "type" : "boolean"
          },
          "pcPower" : {
            "type" : "string"
          },
          "hasXboxGamePass" : {
            "type" : "boolean",
            "nullable" : true
          },
          "preferredGameGenres" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "challengeLevelPreference" : {
            "type" : "boolean",
            "nullable" : true
          },
          "gameLengthPreference" : {
            "type" : "boolean",
            "nullable" : true
          },
          "competitiveBlitzParticipation" : {
            "type" : "boolean"
          },
          "casualBlitzParticipation" : {
            "type" : "boolean"
          },
          "bookClubParticipation" : {
            "type" : "boolean"
          },
          "gameClubParticipation" : {
            "type" : "boolean"
          },
          "notesForStaff" : {
            "type" : "string"
          },
          "notesForContractor" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "access_token" : {
        "name" : "Authorization",
        "type" : "apiKey",
        "in" : "header"
      }
    }
  },
  "security" : [ {
    "access_token" : [ ]
  } ]
}